refactoring: moving markdown parsing to utils
